sourceforge code tidys: 1. Replace use of & with && when working with boolean values 2. Close IOStreams along all paths 3. Actually throw the JS exception in CMClass 4. Remove String comparision using != 5. ints cannot be greater than Integer.MAX_VALUE. Convert to long for computation, then check. 6. incorrect bitmask in com.planet_ink.coffee_mud.Libraries.Sense line 313 and Line 341
Remove unnecessary object creation and function calls:
1. Replace new Boolean() with Boolean.valueOf(), Boolean.TRUE, or
Boolean.FALSE
2. Replace new String([String]) with [String]
3. Replace new String("String") with "String"
5. Remove unnecessary String.toString() calls
6. Remove unnecessary boxing-then-unboxing
7. Remove unnecessary calls to Math.round() for integer values
8. Remove variable masking in subclasses
9. Remove superfluous delegation to superclass
10. Replace new Integer(x).toString() with Integer.toString( x )
11. Replace new Long(x).toString() with Long.toString( x )
12. Replace new Character(x).toString() with Character.toString( x )


git-svn-id: svn://192.168.1.10/public/CoffeeMud@7225 0d6f1817-ed0e-0410-87c9-987e46238f29

+++ com/planet_ink/coffee_mud/Abilities/Spells/Spell_Fatigue.java
@@ -16,7 +16,7 @@ import com.planet_ink.coffee_mud.Races.interfaces.*;
 
 import java.util.*;
 
-/* 
+/*
    Copyright 2000-2008 Bo Zimmerman
 
    Licensed under the Apache License, Version 2.0 (the "License");
@@ -66,7 +66,7 @@ public class Spell_Fatigue extends Spell
 				mob.location().send(mob,msg);
 				if(msg.value()<=0)
 				{
-					target.curState().adjFatigue(CharState.FATIGUED_MILLIS+(mob.envStats().level() * 5 * Tickable.TIME_TICK),target.maxState());
+					target.curState().adjFatigue(CharState.FATIGUED_MILLIS+(((long)mob.envStats().level()) * 5l * Tickable.TIME_TICK),target.maxState());
 					mob.location().show(target,null,CMMsg.MSG_OK_VISUAL,"<S-NAME> feel(s) incredibly fatigued!");
 				}
 			}
